@@ -59,9 +60,9 @@
     "prop-types": "^15.7.2",
     "query-string": "^5.1.1",
     "ramda": "^0.26.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-final-form": "^4.1.0",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "react-final-form": "^6.1.0",
     "react-helmet": "^6.1.0",
     "react-input-range": "^1.3.0",
     "react-router-dom": "^5.2.0",
