@@ -2,105 +2,100 @@ import React from 'react'
 
 export default () => (
   <div className="confined-container">
-    <p className="bottom-spacing-xl">
-      <br />
-      <h3>GDPR</h3>
-    </p>
-    <p>
-      <b>Декларация за защита на личните данни</b>
-      <br />
-      Защитата на личните Ви данни ни е важна. Затова обработваме Вашата
-      информация изключително въз основа на приложимото законодателство,
-      конкретно&nbsp;GDPR, Закона за защита на лични данни (ЗЗЛД) и Закона за
-      електронната търговия (ЗЕТ). В тази декларация ви информираме за
-      най-важните аспекти от обработването на Вашите данни на нашия уебсайт.
-      <br />
-      Какво се случва с данните Ви, след като се свържете се нас?
-      <br />
-      Когато ни пращате имейл, ние съхраняваме данните Ви за период от един
-      месец, така че да можем да обработим искането Ви, както и за да отговорим
-      на Ваши последващи въпроси. Няма да споделяме тази информация с трети лица
-      без Вашето изрично съгласие!
-      <br />
-      <b>Съхраняване на Вашите данни</b>
-      <br />
-      Обръщаме внимание, че с цел облекчаване на процеса по поръчка и покупка,
-      както и последващото изпълнение на договора с Вас, съхраняваме, както и
-      Вашите имена, адрес, телефонен номер, имейл адрес.
-      <br />
-      Предоставените от Вас данни са ни нужни, за да изпълним своите договорни
-      или преддоговорни задължения спрямо Вас. Не предоставяме тези данни на
-      трети лица, с изключение на:
-      <br />
-      &ndash; куриерската фирма, с която Ви доставяме покупката и<br />
-      &ndash; нашия счетоводител.
-      <br />
-      При незавършена поръчка изтриваме въведените по нея данни. При завършена
-      поръчка и изпълнен от наша страна договор на основание чл. 6, ал. 1, букви
-      (а) и (б) от GDPR съхраняваме всички данни по правоотношението до изтичане
-      на предвидения в чл. 12 от Закона за счетоводството (ЗСч) 10-годишен срок.
-      <br />
-      <b>Бисквитки</b>
-      <br />
-      Нашият уеб сайт използва така наречените &bdquo;бисквитки&ldquo;. Това са
-      малки текстови файлове, които се зареждат в браузъра и се съхраняват на
-      Вашето крайно устройство. Те са безобидни.
-      <br />
-      Използваме ги, за да поддържаме сайта си лесен за употреба. Някои
-      &bdquo;бисквитки&ldquo; остават съхранени на устройството Ви, докато не ги
-      изтриете. Те ни позволяват да разпознаем Вашия браузър при следващото ви
-      посещение в нашия сайт.
-      <br />
-      Използваме &bdquo;бисквитките&ldquo; на основание чл. 4а от Закона за
-      електронната търговия (ЗЕТ) и&nbsp;член 6, ал. 1, буква (е)&nbsp;GDPR.
-      <br />
-      Ако не сте съгласни с това, можете да откажете съхраняването, като
-      настроите браузъра си така, че да Ви информира, когато някой сайт иска да
-      запамети на устройството Ви &bdquo;бисквитки&ldquo;, а Вие съответно да ги
-      приемате или не.
-      <br />
-      Имайте предвид, че деактивирането на &bdquo;бисквитките&ldquo; във Вашия
-      браузър може да ограничи функционалността на нашия сайт за Вас.
-      <br />
-      <b>Анализ на интернет трафика</b>
-      <br />
-      Нашият сайт анализира потребителския интернет трафик чрез Google Analytics
-      <br />
-      Обработването на данните за анализ на интернет трафика се извършва въз
-      основа на член 6, ал. 1, буква (a) и / или (е) GDPR.
-      <br />
-      Легитимният ни интерес по смисъла на GDPR се състои в подобряване
-      функционалността на нашия сайт. Ценим личната неприкосновеност на
-      потребителите си, поради което данните, които използваме за анализ на уеб
-      трафика, биват анонимизирани от страна на доставчика на техническото
-      решение. С последния имаме договор за обработка на лични данни.
-      <br />
-      Данните съхраняваме за период от един месец
-      <br />
-      <b>Бюлетин</b>
-      <br />
-      Можете да се абонирате за нашия бюлетин, като попълните формуляра на нашия
-      сайт. За тази цел се нуждаем от Вашия имейл адрес и съгласието Ви да Ви
-      пращаме съответните съобщения.
-      <br />
-      Можете да се откажете от абонамента си за нашия бюлетин по всяко време.
-      Моля, изпратете отказа си на следния електронен адрес:
-      support@damskichanti.com. Веднага ще заличим данните Ви във връзка с
-      изпращането на нашия нюзлетър/мейлинг.
-      <br />
-      <b>Вашите права</b>
-      <br />
-      Имате право на достъп до Вашите данни, които обработваме, както и право на
-      коригиране, изтриване (право &bdquo;да бъдете забравени&ldquo;),
-      ограничаване на обработката, преносимост на данни, оттегляне на съгласието
-      и възражение.
-      <br />
-      Ако смятате, че нашето обработване на Вашите данни нарушава разпоредбите
-      за защита на лични данни или ако смятате, че правата Ви според тези
-      разпоредби са били нарушени по някакъв начин, можете да се обърнете към
-      компетентния административен орган. В България това е&nbsp;Комисията за
-      защита на лични данни.&nbsp;
-      <br />
-    </p>
+    <h3 className="bottom-spacing-xl">GDPR</h3>
+    <b>Декларация за защита на личните данни</b>
+    <br />
+    Защитата на личните Ви данни ни е важна. Затова обработваме Вашата
+    информация изключително въз основа на приложимото законодателство,
+    конкретно&nbsp;GDPR, Закона за защита на лични данни (ЗЗЛД) и Закона за
+    електронната търговия (ЗЕТ). В тази декларация ви информираме за най-важните
+    аспекти от обработването на Вашите данни на нашия уебсайт.
+    <br />
+    Какво се случва с данните Ви, след като се свържете се нас?
+    <br />
+    Когато ни пращате имейл, ние съхраняваме данните Ви за период от един месец,
+    така че да можем да обработим искането Ви, както и за да отговорим на Ваши
+    последващи въпроси. Няма да споделяме тази информация с трети лица без
+    Вашето изрично съгласие!
+    <br />
+    <b>Съхраняване на Вашите данни</b>
+    <br />
+    Обръщаме внимание, че с цел облекчаване на процеса по поръчка и покупка,
+    както и последващото изпълнение на договора с Вас, съхраняваме, както и
+    Вашите имена, адрес, телефонен номер, имейл адрес.
+    <br />
+    Предоставените от Вас данни са ни нужни, за да изпълним своите договорни или
+    преддоговорни задължения спрямо Вас. Не предоставяме тези данни на трети
+    лица, с изключение на:
+    <br />
+    &ndash; куриерската фирма, с която Ви доставяме покупката и<br />
+    &ndash; нашия счетоводител.
+    <br />
+    При незавършена поръчка изтриваме въведените по нея данни. При завършена
+    поръчка и изпълнен от наша страна договор на основание чл. 6, ал. 1, букви
+    (а) и (б) от GDPR съхраняваме всички данни по правоотношението до изтичане
+    на предвидения в чл. 12 от Закона за счетоводството (ЗСч) 10-годишен срок.
+    <br />
+    <b>Бисквитки</b>
+    <br />
+    Нашият уеб сайт използва така наречените &bdquo;бисквитки&ldquo;. Това са
+    малки текстови файлове, които се зареждат в браузъра и се съхраняват на
+    Вашето крайно устройство. Те са безобидни.
+    <br />
+    Използваме ги, за да поддържаме сайта си лесен за употреба. Някои
+    &bdquo;бисквитки&ldquo; остават съхранени на устройството Ви, докато не ги
+    изтриете. Те ни позволяват да разпознаем Вашия браузър при следващото ви
+    посещение в нашия сайт.
+    <br />
+    Използваме &bdquo;бисквитките&ldquo; на основание чл. 4а от Закона за
+    електронната търговия (ЗЕТ) и&nbsp;член 6, ал. 1, буква (е)&nbsp;GDPR.
+    <br />
+    Ако не сте съгласни с това, можете да откажете съхраняването, като настроите
+    браузъра си така, че да Ви информира, когато някой сайт иска да запамети на
+    устройството Ви &bdquo;бисквитки&ldquo;, а Вие съответно да ги приемате или
+    не.
+    <br />
+    Имайте предвид, че деактивирането на &bdquo;бисквитките&ldquo; във Вашия
+    браузър може да ограничи функционалността на нашия сайт за Вас.
+    <br />
+    <b>Анализ на интернет трафика</b>
+    <br />
+    Нашият сайт анализира потребителския интернет трафик чрез Google Analytics
+    <br />
+    Обработването на данните за анализ на интернет трафика се извършва въз
+    основа на член 6, ал. 1, буква (a) и / или (е) GDPR.
+    <br />
+    Легитимният ни интерес по смисъла на GDPR се състои в подобряване
+    функционалността на нашия сайт. Ценим личната неприкосновеност на
+    потребителите си, поради което данните, които използваме за анализ на уеб
+    трафика, биват анонимизирани от страна на доставчика на техническото
+    решение. С последния имаме договор за обработка на лични данни.
+    <br />
+    Данните съхраняваме за период от един месец
+    <br />
+    <b>Бюлетин</b>
+    <br />
+    Можете да се абонирате за нашия бюлетин, като попълните формуляра на нашия
+    сайт. За тази цел се нуждаем от Вашия имейл адрес и съгласието Ви да Ви
+    пращаме съответните съобщения.
+    <br />
+    Можете да се откажете от абонамента си за нашия бюлетин по всяко време.
+    Моля, изпратете отказа си на следния електронен адрес:
+    support@damskichanti.com. Веднага ще заличим данните Ви във връзка с
+    изпращането на нашия нюзлетър/мейлинг.
+    <br />
+    <b>Вашите права</b>
+    <br />
+    Имате право на достъп до Вашите данни, които обработваме, както и право на
+    коригиране, изтриване (право &bdquo;да бъдете забравени&ldquo;),
+    ограничаване на обработката, преносимост на данни, оттегляне на съгласието и
+    възражение.
+    <br />
+    Ако смятате, че нашето обработване на Вашите данни нарушава разпоредбите за
+    защита на лични данни или ако смятате, че правата Ви според тези разпоредби
+    са били нарушени по някакъв начин, можете да се обърнете към компетентния
+    административен орган. В България това е&nbsp;Комисията за защита на лични
+    данни.&nbsp;
+    <br />
   </div>
 )
