@@ -6,56 +6,26 @@ const ObjectID = require('mongodb').ObjectID
 const { sign, verify } = require('jsonwebtoken')
 const { secret } = require('../../.config.json')
 
-const mergeCarts = async (user, currentCartId, cartService, db) => {
-  const userCartId = user.cart || currentCartId
-  if (!user.cart) {
-    db.collection('users').updateOne(
-      {
-        email: user.email
-      },
-      {
-        $set: { cart: userCartId }
-      }
-    )
-  }
-  if (userCartId !== currentCartId) {
-    const userDbProducts = await cartService.getCartItems(userCartId)
-    const currentCartProducts = await cartService.getCartItems(currentCartId)
-    const cartProducts = cartService.mergeCartProducts(
-      currentCartProducts,
-      userDbProducts
-    )
-    if (!R.equals(currentCartProducts, userDbProducts)) {
-      await db.collection('carts').updateOne(
-        {
-          _id: ObjectID(userCartId)
-        },
-        {
-          $set: {
-            products: cartProducts
-          }
-        }
-      )
-    }
-    cartService.clearCart(currentCartId)
-  }
-
-  return userCartId
-}
-const verifyUser = async (password, { salt, password: hashPassword }) => {
+const verifyUser = (password, { salt, password: hashPassword }) => {
   if (salt) {
     const result = sha1(concat(salt, sha1(concat(salt, sha1(password)))))
     return result === hashPassword
   }
 
-  return await bcrypt.compare(password, hashPassword)
+  return bcrypt.compare(password, hashPassword)
+}
+const createResetPassToken = resetPassEmail => {
+  return sign({ resetPassEmail }, secret, { expiresIn: '24h' })
 }
 module.exports = (db, cartService) => ({
   verify: token => {
     return verify(token, secret)
   },
-  register: async ({ email, password, name, lastname, consent, cart }) => {
-    const hashedPassword = await bcrypt.hash(password, 1)
+  hashPassword: function(password) {
+    return bcrypt.hash(password, 1)
+  },
+  register: async ({ email, password, name, lastname, consent }) => {
+    const hashedPassword = await this.hashPassword(password)
     const newCart = await cartService.createNewCart()
     const result = await db.collection('users').insertOne({
       email,
@@ -66,8 +36,12 @@ module.exports = (db, cartService) => ({
       consent
     })
     const user = result.ops[0]
-    const userCart = await mergeCarts(user, cart, cartService, db)
-    user.cart = userCart
+    await cartService.mergeCarts(user.cart, req.cart)
+
+    res.cookie('cart', user.cart, {
+      maxAge: 86400 * 30 * 1000,
+      httpOnly: true
+    })
 
     return {
       token: sign({ email: user.email }, secret, {
@@ -76,13 +50,19 @@ module.exports = (db, cartService) => ({
       user
     }
   },
-  login: async (email, password, currentCartId) => {
+  createResetPassToken,
+  login: async (email, password) => {
     const user = await db.collection('users').findOne({
       email
     })
+
+    if(user === null) {
+      return null
+    }
+
     const verified = await verifyUser(password, user)
+
     if (verified) {
-      user.cart = await mergeCarts(user, currentCartId, cartService, db)
       return {
         token: sign({ email: user.email }, secret, {
           expiresIn: '6h'
