@@ -1,4 +1,5 @@
 const ObjectID = require('mongodb').ObjectID
+
 module.exports = (cartProvider, db, currentCartId) => ({
   getCustomerCart: cartProvider.getCustomerCart,
   createCartFromAdaptedRecords: cartProvider.adaptedRecordsToCart,
@@ -7,9 +8,9 @@ module.exports = (cartProvider, db, currentCartId) => ({
   },
   createCart: records => cartProvider.createCart(records),
   mergeCartProducts: cartProvider.mergeCartProducts,
-  getCartItems: async cartId => {
-    const { products } = await cartProvider.getDbCart(cartId)
-    return products
+  getCartItems: async function(cartId) {
+    const data = await cartProvider.getDbCart(cartId)
+    return data ? data.products : []
   },
   clearCart: cartProvider.clearCart,
   createNewCart: async () => {
@@ -43,5 +44,29 @@ module.exports = (cartProvider, db, currentCartId) => ({
         }
       }
     )
+  },
+  mergeCarts: async function(userCartId, currentCartId) {
+    if (userCartId !== currentCartId) {
+      const userDbProducts = await this.getCartItems(userCartId)
+      const currentCartProducts = await this.getCartItems(currentCartId)
+      const cartProducts = cartProvider.mergeCartProducts(
+        currentCartProducts,
+        userDbProducts
+      )
+      if (!R.equals(currentCartProducts, userDbProducts)) {
+        await db.collection('carts').updateOne(
+          {
+            _id: ObjectID(userCartId)
+          },
+          {
+            $set: {
+              products: cartProducts
+            }
+          }
+        )
+      }
+      cartService.clearCart(currentCartId)
+    }
+    return userCartId
   }
 })
